fix(测试跟踪): 用例库列表标签显示优化
--bug=1023677 --user=陈建星 【测试跟踪】功能测试-公共用例库列表-标签为空显示为[] https://www.tapd.cn/55049933/s/1340774
